feat: finish most of apis
