+++ lib/config.js
@@ -2,6 +2,7 @@
 
 const HOST_RELEASE = 'https://press.one';
 const HOST_DEV = 'https://beta.press.one';
+// const HOST_DEV = '127.0.0.1:8090';
 
 module.exports = PRSConfig;
