+++ lib/index.js
@@ -9,6 +9,7 @@ const File = require('./file');
 const Block = require('./block');
 const Draft = require('./draft');
 const Contract = require('./contract');
+const Order = require('./order');
 const DApp = require('./dapp');
 const signUtility = require('./signUtility');
 const { hashRequest, signRequest, getAuthHeader } = require('./request');
@@ -36,6 +37,7 @@ function PRS (options) {
   this.block = new Block(this.config);
   this.draft = new Draft(this.config);
   this.contract = new Contract(this.config);
+  this.order = new Order(this.config);
   this.dapp = new DApp(this.config);
 }
